(fix): Fixed bugs that prevented it from running
